Refactoring the Location Methods

+++ lib/screens/location_screen.dart
@@ -27,6 +27,13 @@ class _LocationScreenState extends State<LocationScreen> {
 
   void updateUI(Map<String, dynamic> weatherData) {
     setState(() {
+      if (weatherData == null) {
+        temperature = 0;
+        weatherIcon = 'Error';
+        weatherMessage = 'Unable to get the weather data';
+        cityName = '';
+        return;
+      }
       double temp = weatherData['main']['temp'] + 0.0;
       temperature = temp.truncate();
       weatherMessage = weather.getMessage(temperature);
@@ -62,7 +69,12 @@ class _LocationScreenState extends State<LocationScreen> {
                 mainAxisAlignment: MainAxisAlignment.spaceBetween,
                 children: <Widget>[
                   FlatButton(
-                    onPressed: () {},
+                    onPressed: () async {
+                      Map<String, dynamic> weatherData =
+                          await weather.getLocationWeather();
+
+                      updateUI(weatherData);
+                    },
                     child: Icon(
                       Icons.near_me,
                       size: 48.0,
